refactor: address PR comments
