+++ packages/dom/src/utilities/frame/get-frame-element.ts
@@ -1,0 +1,9 @@
+export function getFrameElement(el: Element) {
+  const refWindow = el.ownerDocument.defaultView;
+
+  if (refWindow && refWindow.self !== refWindow.parent) {
+    return refWindow.frameElement;
+  }
+
+  return null;
+}
