--- packages/dom/src/utilities/iframe/get-iframe.ts
@@ -1,11 +1,0 @@
-export default function getIframe(el: Element) {
-  const refWindow = el.ownerDocument.defaultView;
-
-  if (refWindow && refWindow.self !== refWindow.parent) {
-    const iframe = refWindow.frameElement as HTMLIFrameElement | null;
-
-    return iframe;
-  }
-
-  return null;
-}
