+++ packages/dom/src/utilities/index.ts
@@ -49,3 +49,6 @@ export {
 export type {Transform} from './transform/index.ts';
 
 export {generateUniqueId} from './misc/generateUniqueId.ts';
+
+export {getFrameElement} from './frame/get-frame-element.ts';
+export {getFrameOffset} from './frame/get-frame-offset.ts';
